# docs/swagger.yaml
definitions:
  auth_request.AuthRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  auth_request.CBTAuthRequest:
    properties:
      username:
        type: string
    type: object
  auth_response.AuthResponse:
    properties:
      detail: {}
      exp:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_Sistem-Informasi-Akademik_academic-system-information-service_src_main_model_entity_user.User'
    type: object
  auth_response.AuthResponseCBT:
    properties:
      exam:
        $ref: '#/definitions/school.Exam'
      exam_session:
        $ref: '#/definitions/school.ExamSession'
      exam_taken:
        $ref: '#/definitions/cbt.StudentHistoryTaken'
      exp:
        type: integer
      token:
        type: string
      user:
        $ref: '#/definitions/student.StudentClass'
    type: object
  cbt.StudentHistoryTaken:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      end_at:
        type: string
      exam_code:
        type: string
      id:
        type: integer
      is_cheating:
        type: boolean
      is_finished:
        type: boolean
      is_forced:
        type: boolean
      is_time_over:
        type: boolean
      need_correction:
        type: boolean
      remaining_time:
        type: integer
      score:
        type: integer
      session_id:
        description: DetailExam    school.Exam        `json:"detail_exam" gorm:"foreignKey:ExamCode;references:ExamCode"`
        type: string
      start_at:
        description: DetailStudent student.Student    `json:"detail_student" gorm:"foreignKey:StudentId;references:ID"`
        type: string
      status:
        type: string
      student_id:
        description: DetailSession school.ExamSession `json:"detail_session" gorm:"foreignKey:SessionId;references:SessionId"`
        type: integer
      total_correct:
        type: integer
      total_wrong:
        type: integer
      updatedAt:
        type: string
    type: object
...
  class_response.ClassCodeResponse:
    properties:
      code:
        type: string
      name:
        type: string
    type: object
...
  curriculum.Subject:
    properties:
      code:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      subject:
        type: string
      subject_type:
        type: string
      updatedAt:
        type: string
    type: object
...
  github_com_Sistem-Informasi-Akademik_academic-system-information-service_src_main_model_entity_user.User:
    properties:
      SchoolCode:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/user.Role'
      status:
        type: integer
      updatedAt:
        type: string
      username:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  response.BaseResponse:
    properties:
      code:
        type: integer
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: string
    type: object
...
  school.Class:
    properties:
      classCode:
        type: string
      className:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      detailClassCode:
        $ref: '#/definitions/school.ClassCode'
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  school.ClassCode:
    properties:
      class_member:
        items:
          $ref: '#/definitions/school.Class'
        type: array
      code:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
  school.Exam:
    properties:
      code:
        type: string
      created_by:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      detail_type_exam:
        $ref: '#/definitions/school.TypeExam'
      duration:
        type: integer
      exam_member:
        items:
          $ref: '#/definitions/school.ExamMember'
        type: array
      exam_question:
        items:
          $ref: '#/definitions/school.ExamQuestion'
        type: array
      id:
        type: integer
      modified_by:
        description: CreatedUser  user.User `json:"detail_user" gorm:"foreignKey:CreatedBy"`
        type: integer
      name:
        type: string
      random_answer:
        type: boolean
      random_question:
        type: boolean
      show_result:
        type: boolean
      subject_code:
        $ref: '#/definitions/curriculum.Subject'
      total_score:
        type: integer
      type_question:
        type: string
      updatedAt:
        type: string
    type: object
  school.ExamAnswerOption:
    properties:
      answer_id:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      option:
        type: string
      question_id:
        type: string
      updatedAt:
        type: string
    type: object
  school.ExamMember:
    properties:
      class:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      detail_class:
        $ref: '#/definitions/school.Class'
      detail_exam:
        $ref: '#/definitions/school.Exam'
      exam_code:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
  school.ExamQuestion:
    properties:
      answer_single:
        type: string
      bank_question_id:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      exam_code:
        type: string
      id:
        type: integer
      question:
        type: string
      question_from:
        type: string
      question_id:
        type: string
      question_option:
        items:
          $ref: '#/definitions/school.ExamAnswerOption'
        type: array
      score:
        type: integer
      type_question:
        type: string
      updatedAt:
        type: string
    type: object
  school.ExamSession:
    properties:
      created_by:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      detail_exam:
        $ref: '#/definitions/school.Exam'
      end_date:
        type: string
      error_report:
        type: string
      exam_member:
        items:
          $ref: '#/definitions/school.ExamSessionMember'
        type: array
      id:
        type: integer
      modified_by:
        description: CreatedUser  user.User `json:"detail_user" gorm:"foreignKey:CreatedBy"`
        type: integer
      name:
        type: string
      report_url:
        type: string
      session_id:
        type: string
      start_date:
        type: string
      status_report:
        type: string
      updatedAt:
        type: string
    type: object
  school.ExamSessionMember:
    properties:
      class:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      detail_class:
        $ref: '#/definitions/school.Class'
      detail_exam:
        $ref: '#/definitions/school.ExamSession'
      id:
        type: integer
      session_id:
        type: string
      updatedAt:
        type: string
    type: object
...
  school.TypeExam:
    properties:
      code:
        type: string
      color:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      detail_role:
        $ref: '#/definitions/user.Role'
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  school_request.ModifySchoolRequest:
    properties:
      address:
        type: string
      banner:
        type: string
      email:
        type: string
      level_of_education:
        type: string
      logo:
        type: string
      npsn:
        type: string
      nss:
        type: string
      phone:
        type: string
      principal_name:
        type: string
      principal_nip:
        type: string
      school_name:
        type: string
      vice_principal_name:
        type: string
      vice_principal_nip:
        type: string
    type: object
  school_response.DashboardResponse:
    properties:
      total_access:
        type: integer
      total_class:
        type: integer
      total_exam:
        type: integer
      total_report_exam:
        type: integer
      total_session_exam:
        type: integer
      total_student:
        type: integer
      total_subject:
        type: integer
    type: object
  school_response.DetailSchool:
    properties:
      address:
        type: string
      banner:
        type: string
      email:
        type: string
      id:
        type: string
      level_of_education:
        type: string
      logo:
        type: string
      npsn:
        type: string
      nss:
        type: string
      phone:
        type: string
      principal_name:
        type: string
      principal_nip:
        type: string
      school_name:
        type: string
      vice_principal_name:
        type: string
      vice_principal_nip:
        type: string
    type: object
  student.Student:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      detail_user:
        $ref: '#/definitions/github_com_Sistem-Informasi-Akademik_academic-system-information-service_src_main_model_entity_user.User'
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      nisn:
        type: string
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  student.StudentClass:
    properties:
      class_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      detail_class:
        $ref: '#/definitions/school.Class'
      detail_student:
        $ref: '#/definitions/student.Student'
      id:
        type: integer
      student_id:
        type: integer
      updatedAt:
        type: string
    type: object
...
  user.Role:
    properties:
      SchoolCode:
        type: string
      code:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
...
host: api.sman16bekasi.id/gateway-service
info:
  contact:
    email: cbt@sman16bekasi.id
    name: CBT Team
    url: https://sman16bekasi.id
  description: This is documentation technical
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: CBT Documentation
  version: "1.0"
paths:
  /academic/auth/login:
    post:
      consumes:
      - application/json
      description: Auth login for admin
      parameters:
      - description: Body Request of Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth_request.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login Response
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth_response.AuthResponse'
              type: object
      summary: This endpoint about auth login
      tags:
      - Auth
  /academic/class-code:
    get:
      consumes:
      - application/json
      description: Return Class Code list data
      produces:
      - application/json
      responses:
        "200":
          description: Class Code Response
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/class_response.ClassCodeResponse'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: This endpoint about Class Code like 10, 11, 12
      tags:
      - School
...
  /academic/dashboard:
    get:
      consumes:
      - application/json
      description: Return dashboard user
      produces:
      - application/json
      responses:
        "200":
          description: Dashboard response
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/school_response.DashboardResponse'
              type: object
      security:
      - BearerAuth: []
      summary: This endpoint about dashboard user
      tags:
      - School
...
  /academic/school:
    get:
      consumes:
      - application/json
      description: Return school data information
      parameters:
      - description: Required school code
        in: query
        name: schoolCode
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: School information
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/school_response.DetailSchool'
              type: object
      security: []
      summary: This endpoint about school information
      tags:
      - School
  /academic/school/update:
    put:
      consumes:
      - application/json
      description: Update data school
      parameters:
      - description: Body Request of Update school
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/school_request.ModifySchoolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: School detail Response
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/school_response.DetailSchool'
              type: object
      security:
      - BearerAuth: []
      summary: This endpoint Update data school
      tags:
      - School
...
  /academic/subjects:
    get:
      consumes:
      - application/json
      description: Return Subject list data
      produces:
      - application/json
      responses:
        "200":
          description: subjects Response
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/curriculum.Subject'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: This endpoint get data all subjects like Matematika and others
      tags:
      - School
...
  /auth/cbt/login:
    post:
      consumes:
      - application/json
      description: Auth login for Student
      parameters:
      - description: Body Request of Login
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth_request.CBTAuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login Response
          schema:
            allOf:
            - $ref: '#/definitions/response.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth_response.AuthResponseCBT'
              type: object
      summary: This endpoint about auth login
      tags:
      - Auth
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Use "Bearer {your_token}" to authenticate. Some endpoints are public
      and do not require authentication.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...